Reshuffle into ProtoCompat/FromProto
